Refractored the Search Screen and added it to the main and added a few comments
