+++ src/main/java/APP_Business_Rules/login_user/LoginUserResponseModel.java
@@ -4,12 +4,19 @@ public class LoginUserResponseModel {
 
     LoginUserGatewayModel model;
 
-
+    /**
+     * The result for when a user has attempted a login.
+     * @param model the information found from a login.
+     */
     public LoginUserResponseModel(LoginUserGatewayModel model) {
         this.model = model;
     }
 
     public String getUsername(){
         return this.model.getUsername();
+    }
+
+    public LoginUserGatewayModel getModel(){
+        return this.model;
     }
 }
