+++ src/main/java/Frameworks_and_Drivers/UI.java
@@ -3,5 +3,6 @@ package Frameworks_and_Drivers;
 public interface UI {
     void updateRestaurantTable(String[][] data);
     void updateDishTable(String[][] data);
-    void updateMenuTable(String[][] data);
+
+    void defaultTable();
 }
