+++ src/main/java/Interface_and_Adapters/Main.java
@@ -2,18 +2,18 @@ package Interface_and_Adapters;
 
 
 import APP_Business_Rules.DishMenu.DishFileReader;
-import APP_Business_Rules.MenuUseCase.MenuInteractor;
 import APP_Business_Rules.RestaurantUseCase.RestaurantFileReader;
 import APP_Business_Rules.SearchUseCase.SearchDishUseCase;
 import APP_Business_Rules.SearchUseCase.SearchRestaurantUseCase;
 import APP_Business_Rules.create_user.*;
-import APP_Business_Rules.login_user.*;
+
 import Entities.AccountFactory;
 import Entities.UserFactory;
-import Frameworks_and_Drivers.AccountUserFile;
+import Frameworks_and_Drivers.SearchScreen;
 import Frameworks_and_Drivers.UserFile;
-import Frameworks_and_Drivers.View;
-import Interface_and_Adapters.StartUpScreens.*;
+import Interface_and_Adapters.StartUpScreens.CreateUserResponse;
+import Interface_and_Adapters.StartUpScreens.SignUpScreen;
+import Interface_and_Adapters.StartUpScreens.StartUpScreen;
 
 import javax.swing.*;
 import java.awt.*;
@@ -21,32 +21,53 @@ import java.awt.*;
 
 public class Main extends JFrame{
 
-    public static void main(String[] args){
-        UserFile userFile = new UserFile("users.csv"); //User file instance
-        AccountUserFile accountUserFile = new AccountUserFile("accounts.csv"); // Account user file instance
-        AccountFactory accountFactory = new AccountFactory(); // Account factory instance
-        LoginUserResponse loginUserResponse = new LoginUserResponse(); // Login user response instance
-        LoginUserInteractor loginUserInteractor = new LoginUserInteractor(userFile, accountUserFile, accountFactory, loginUserResponse); // Login user interactor instance
-        LoginUserController loginUserController = new LoginUserController(loginUserInteractor); // Login user controller instance
-        CreateUserResponse createUserResponse = new CreateUserResponse();
-        CreateUserInteractor createUserInteractor = new CreateUserInteractor(userFile, accountFactory, createUserResponse);
-        CreateUserController createUserController = new CreateUserController(createUserInteractor);
+    public Main() {
+        JPanel mainPanel = new JPanel();
+
+
+        CreateUserGateway user;
+        user = new UserFile("./users.csv");
+        CreateUserPresenter presenter = new CreateUserResponse();
+        UserFactory userFactory = new AccountFactory();
+        CreateUserInputBoundary interactor = new CreateUserInteractor(
+                user, userFactory, presenter);
+        CreateUserController controller = new CreateUserController(interactor);
+
+        //Initializing file readers
+        DishFileReader dishFileReader = new DishFileReader("src/main/java/Frameworks_and_Drivers/dish.csv");
         RestaurantFileReader restaurantFileReader = new RestaurantFileReader("src/main/java/Frameworks_and_Drivers/Restaurant.csv");
-        DishFileReader dishFileReader = new DishFileReader("src/main/java/Frameworks_and_Drivers/Dishes.csv");
+
+        //Initializing classes that contribute to the search Usecase
         SearchPresenter searchPresenter = new SearchPresenter();
-        SearchRestaurantUseCase searchRestaurantUseCase = new SearchRestaurantUseCase(searchPresenter,restaurantFileReader);
         SearchDishUseCase searchDishUseCase = new SearchDishUseCase(searchPresenter, dishFileReader);
+        SearchRestaurantUseCase searchRestaurantUseCase = new SearchRestaurantUseCase(searchPresenter, restaurantFileReader);
         SearchController searchController = new SearchController(searchDishUseCase, searchRestaurantUseCase);
-        MenuPresenter menuPresenter = new MenuPresenter();
-        MenuInteractor menuInteractor = new MenuInteractor(dishFileReader, menuPresenter);
-        MenuController menuController = new MenuController(menuInteractor);
-        View view = new View(loginUserController, createUserController, searchController, menuController);
-        menuPresenter.setView(view);
-        searchPresenter.setView(view);
-        JFrame frame = new JFrame("view");
-        frame.setContentPane(view.getMainPanel());
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        frame.pack();
-        frame.setVisible(true);
+        SearchScreen searchScreen = new SearchScreen(searchController, mainPanel);
+        searchPresenter.setView(searchScreen);
+        //
+
+        this.setTitle("Digital Dining Divas");
+        this.setResizable(false);
+        this.setExtendedState(JFrame.MAXIMIZED_BOTH);
+        this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        this.setLocationRelativeTo(null);
+
+        mainPanel.setLayout(new CardLayout());
+        mainPanel.add(new StartUpScreen(mainPanel), "FIRST");
+        mainPanel.add(new SignUpScreen(controller, mainPanel), "SECOND");
+        mainPanel.add(searchScreen.getResearchPanel(), "FOURTH");
+        this.setContentPane(mainPanel);
+        switchPanel(mainPanel, "FOURTH");
+
+    }
+
+    public void switchPanel(Container container, String panelName) {
+        CardLayout card = (CardLayout) (container.getLayout());
+        card.show(container, panelName);
+    }
+
+    public static void main(String[] args) {
+        new Main().setVisible(true);
+
     }
 }
